MRD-2987 Move types to appropriate packages
The domain/prisonapi package should contain the types of the responses we get
from the Prison API. The domain/makerecalldecisions/prison package should
contain the types we use internally in our application. Converters should be
used to map between these two sets of types.
